foillogic headers to include dir
