+++ src/foileditors/outlineeditor/outlineeditor.h
@@ -24,7 +24,7 @@
 #define OUTLINEEDITOR_H
 
 #include "patheditor/fwd/patheditorfwd.hpp"
-#include "foillogicfwd/foillogicfwd.h"
+#include "foillogic/fwd/foillogicfwd.hpp"
 
 #include <QWidget>
 #include <memory>
